The imei is now saved in the correct place
