+++ app/src/main/java/com/apptec/registrateapp/repository/localDatabase/DAOs/DeviceDao.java
@@ -10,6 +10,7 @@ import com.apptec.registrateapp.models.Device;
 
 import java.util.List;
 
+
 @Dao
 public interface DeviceDao {
     /**
