+++ app/src/main/java/com/apptec/registrateapp/view/fragments/device2/DeviceFragment22.java
@@ -17,7 +17,7 @@ import com.apptec.registrateapp.R;
 import com.apptec.registrateapp.models.Device;
 import com.apptec.registrateapp.presenter.DevicePresenterImpl;
 import com.apptec.registrateapp.repository.localDatabase.DatabaseAdapter;
-import com.apptec.registrateapp.view.modals.DialogDevice;
+import com.apptec.registrateapp.view.modals.DialogDevice2;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import java.util.ArrayList;
@@ -80,7 +80,7 @@ public class DeviceFragment22 extends Fragment implements DeviceView2 {
         fabAddDevice.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                DialogDevice df = new DialogDevice();
+                DialogDevice2 df = new DialogDevice2();
                 df.show(getFragmentManager(),"DialogDevice");
             }
         });
